update route and add test
